move default options functions to a new file
